Refactor visualizerMachine file
