+++ src/machines/visualizerMachine/visualizerMachine.ts
@@ -1,5 +1,3 @@
-import { MouseEventHandler } from "react";
-import { Options as RoughJSOptions } from "roughjs/bin/core";
 import invariant from "tiny-invariant";
 import { v4 as uuidv4 } from "uuid";
 import { assign, createMachine } from "xstate";
@@ -10,7 +8,6 @@ import {
   calculateNormalizedValue,
   calculateNormalizedZoom,
   isIntersecting,
-  Point,
   createElement,
   isGenericElement,
   isLinearElement,
@@ -19,258 +16,15 @@ import {
   isFreeDrawElement,
   calculateClientPoint,
   measureText,
-} from "../utils";
+} from "../../utils";
 import debounce from "lodash.debounce";
-import { TEXTAREA_UNIT_LESS_LINE_HEIGHT } from "../constants";
-
-export type VisualizerElementBase = {
-  id: string;
-
-  // VisualizerGenericElement: left top point
-  // VisualizerLinearElement: start point
-  x: number;
-  y: number;
-
-  // VisualizerGenericElement: absolute width, height
-  // VisualizerLinearElement: absolute width, height
-  width: number;
-  height: number;
-
-  isSelected: boolean;
-  options: RoughJSOptions;
-  isDeleted: boolean;
-};
-
-type VisualizerSelectionElement = VisualizerElementBase & {
-  shape: "selection";
-};
-type VisualizerRectangleElement = VisualizerElementBase & {
-  shape: "rectangle";
-};
-type VisualizerEllipseElement = VisualizerElementBase & {
-  shape: "ellipse";
-};
-
-export type VisualizerGenericElement =
-  | VisualizerSelectionElement
-  | VisualizerRectangleElement
-  | VisualizerEllipseElement;
-
-type TuplePoint = [number, number];
-
-export type VisualizerLinearElement = VisualizerElementBase & {
-  shape: "line" | "arrow";
-  points: TuplePoint[];
-};
-
-export type VisualizerFreeDrawElement = VisualizerElementBase & {
-  shape: "freedraw";
-  points: TuplePoint[];
-};
-
-export type FontSize = 12 | 16 | 20 | 24;
-type VisualizerTextElementOptions = {
-  fontFamily: "serif";
-  fontSize: FontSize;
-};
-
-export type VisualizerTextElement = VisualizerElementBase &
-  VisualizerTextElementOptions & {
-    shape: "text";
-    text: string;
-  };
-
-export type VisualizerElement =
-  | VisualizerGenericElement
-  | VisualizerLinearElement
-  | VisualizerFreeDrawElement
-  | VisualizerTextElement;
-
-export const SHAPE_TYPES: Record<
-  VisualizerElement["shape"],
-  "generic" | "linear" | "freedraw" | "text"
-> = {
-  selection: "generic",
-  rectangle: "generic",
-  ellipse: "generic",
-  line: "linear",
-  arrow: "linear",
-  freedraw: "freedraw",
-  text: "text",
-};
-
-type Version = {
-  elements: VisualizerMachineContext["elements"];
-  elementOptions: VisualizerMachineContext["elementOptions"];
-};
-
-type ElementOptions = RoughJSOptions & VisualizerTextElementOptions;
-
-// context that is saved to localStorage
-type VisualizerMachinePersistedContext = {
-  elements: VisualizerElement[];
-  elementOptions: ElementOptions;
-
-  elementShape: VisualizerElement["shape"];
-  drawingElementId: VisualizerElement["id"] | null;
-  drawStartPoint: Point;
-  previousPoint: Point;
-  copiedElements: VisualizerElement[];
-  currentPoint: Point;
-  isElementShapeFixed: boolean;
-  zoom: number;
-  origin: {
-    x: number;
-    y: number;
-  };
-};
-
-export type VisualizerMachineContext = VisualizerMachinePersistedContext & {
-  // context that is **not** saved to localStorage
-  history: Version[];
-  historyStep: number;
-};
-
-/* #region events within machine  */
-export type VisualizerMachineEvents =
-  | {
-      type: "DRAW_START";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "DRAW";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "DRAW_END";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "DELETE_SELECTION";
-    }
-  | {
-      type: "CHANGE_ELEMENT_SHAPE";
-      elementShape: VisualizerElement["shape"];
-    }
-  | {
-      type: "DRAG_START";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "DRAG";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "DRAG_END";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "SELECTED_ELEMENTS.DELETE";
-    }
-  | {
-      type: "SELECTED_ELEMENTS.COPY";
-    }
-  | {
-      type: "SELECTED_ELEMENTS.CUT";
-    }
-  | {
-      type: "SELECTED_ELEMENTS.PASTE";
-      canvasElement: HTMLCanvasElement;
-    }
-  | {
-      type: "MOUSE_MOVE";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "IS_ELEMENT_SHAPE_FIXED_TOGGLE";
-    }
-  | {
-      type: "CHANGE_ELEMENT_OPTIONS";
-      elementOptions: Partial<VisualizerMachineContext["elementOptions"]>;
-    }
-  | {
-      type: "CHANGE_ZOOM";
-      setZoom: (
-        zoom: VisualizerMachineContext["zoom"]
-      ) => VisualizerMachineContext["zoom"];
-      canvasElement: HTMLCanvasElement;
-    }
-  | {
-      type: "CHANGE_ZOOM_WITH_PINCH";
-      setZoom: (
-        zoom: VisualizerMachineContext["zoom"]
-      ) => VisualizerMachineContext["zoom"];
-    }
-  | {
-      type: "PAN";
-      event: WheelEvent;
-      devicePixelRatio: number;
-    }
-  | {
-      type: "HISTORY_UPDATE";
-      changedStep: 1 | -1;
-    }
-  | {
-      type: "WRITE_START";
-      event: Parameters<MouseEventHandler<HTMLCanvasElement>>[0];
-      devicePixelRatio: number;
-    }
-  | {
-      type: "WRITE_END";
-      text: string;
-      canvasElement: HTMLCanvasElement;
-    };
-/* #endregion */
-
-// zoom in ratio
-export const ZOOM = {
-  DEFAULT: 1,
-  MINIMUM: 0.1,
-  MAXIMUM: 30,
-} as const;
-
-const PERSISTED_CONTEXT: VisualizerMachinePersistedContext = {
-  // track
-  elements: [],
-  elementOptions: {
-    stroke: "#000",
-    fill: "transparent",
-    strokeWidth: 2,
-    fontSize: 16,
-    fontFamily: "serif",
-  },
-
-  // untrack
-  elementShape: "selection",
-  copiedElements: [],
-  drawingElementId: null,
-  drawStartPoint: {
-    x: 0,
-    y: 0,
-  },
-  previousPoint: {
-    x: 0,
-    y: 0,
-  },
-  // cursor point in actual canvas size (not viewport)
-  currentPoint: {
-    x: 0,
-    y: 0,
-  },
-  isElementShapeFixed: false,
-  zoom: ZOOM.DEFAULT,
-  origin: {
-    x: 0,
-    y: 0,
-  },
-};
+import { TEXTAREA_UNIT_LESS_LINE_HEIGHT } from "../../constants";
+import {
+  VisualizerElement,
+  VisualizerMachineContext,
+  VisualizerMachineEvents,
+} from "./types";
+import { PERSISTED_CONTEXT } from "./constant";
 
 export const visualizerMachine =
   /** @xstate-layout N4IgpgJg5mDOIC5QDcCWsCuBDANqgXmAE4AEAtlgMYAWqAdmAHSoQ5gDEAIgEoCCA6gH0AygBVe3UQG0ADAF1EoAA4B7WKgAuqFXUUgAHogDsMo4wCcARgAspgKyXz1hwCYjAGhABPRJYAcMuaMLgBsAMwyYSGOLnYh1uYAvomeaJi4BMTkVLQMzKwcAMIAErwAcgDiAKKCVQAyVQCyVWWiIqUAClWyCkggqupaOnqGCCZmVrZGDk6uHt6ILjIyjNPLMksyzmHmdsmp6Nh4hKQUNPRMLGxcfBUi4pI9egOa2rp9o+MWNvaOzpZuTw+BBLfyMMJGFwuJzLEIBeL7EBpI6ZU45C75a7CepVQqiKqcWoNZqtYSMTg4-FPPovIbvUCfUzfKYzf6AhYIAGWMwuPwAsJhbnLMLWRHIjInbLnPJXDjYhp4glEpotURkwoAeQ6AE1qco1K9hh9jEzJr9ZgD5sCwi5rIxLHFYhEdn47EZLGLDhKsmdcpcCux5bj8YScSS1YwOrwxN15M8DXSRiaJj9pn85kDfIKwqsRWE7NY4XY-NYwn5Peljj70TKA40NQBVbGCesANVjvX1gzeSYQEMsjG2lnCsyMZbCmc55ihq15RkhO3ngr2KSRXqraOl-uuAElhMrw+1eF1BAAxHcADSVog1FQqDT1-QTPeNY1NqdZGY51iWjBCgQhZw7DiEITBXA5K1RKU-UxOUcUVUNiVVdUG2kOMaWfI0GWTZlzTZK1EAcMwZBCW1zH-d1rHnCsUUlX0MVldgSnKaoD1VQQtVEHcNTKYRH1pF9sLfFMWXTS1JxsSwVn8IwqNImRhzsCEaO9TcYMY5jKhqAAtDUNUafjMPpAwcLNNMLXZYF4hWQs-D8fMBQFEIQnAtdILomttyKUotMEXT9MEfgd1EYpBA6HcyhKQzuywkzhNw8z8MnMcQgsFxHDiEsrCMFyVI3aCGIDKMymiw1jMZES8K-YF7JzCE7FMez4nMPxSLyqD6Nra5ij3G9uG1QQGw6TheCpdCuzK3svjMz9xI5acgj8XlHBagIjHIj1V3FfLOq89h+G4YKajECQ0M7J8YvK0yPzEyzfBIlZGpcHZzGnZ6bXajyt1gpifNYsRtQaPjxouybX2mm6LIIzkGtS6y7HMeduULKJPurb6ICILAAHd6CgG4BFKxNXykkVGARpH0pa6Y-AksccyMVqpisZqHDRtSMUxnG8YJoQWk4InBLigFnuCPNGfMHZwksCTSLMHYZDdExAkVhEtvXDrPMYLncbofGeAEWoygFyxzoE2LRhFnNntLCWpcFCSAjtExBWmCJXuiUV1fc9GYJ1nmKQafERHgrieMFi3ECWlZSIapZ5xckwXFl11GFdZzyMlki3fZgq8i5qAoAD24I6uzlIgHUxIhCRGonMBToZiGPpzAnKHRMEJc927WscL4veDufnS97K2xdtvxJfIh2OSWWI-yhOcpNsLYu61nAVCwCAeeH19lvBUD0sieOdknUEJkFPwTG5JXplX76lGIdRYC0PX2B3oT0oiNPlpFSFnAUuwk4+Tk3WNnHKAQJ6d29rRX2GIwBsDIGAOgGgSCwGoFgB+JAiBwDABoN+INzZl2vgOUIwFyKuiksKR2Nc-wAVstOJaBY74wWQI-N4WD4FgCwLASA+CzZGV7J-FYS0+S-1tA1B0EkGo5ksACEs7p2411cttTW31sZEFeK-A6R0jYCwIQIkmUJrbiwnvbGWHIpJjnJgWEsdkXJuDCMkVcdAVAQDgHoFRX0-Txkur2AAtCEScfi7AWFegwpY0drC2BcMwwqbAfFgyEj+FwFgnKyIFIEcIgSZ4bDtP4fwk9LD5msDYWJ+csa6ygAk4mQlLErHdJfFKjVJEWNaikmIsl3SMzLFRMpTAC5Fz1tUoWlts5-m5OlWIS1SyOFPlsIIgQ7GuhdsOPpjB16bzxsMyOCAfwSVejmGZq16bdPLNA1SecmAPyIE-F+VSMK+N3qQ3CdkRS7AUpOcIfhaFRKRl0twazOGIOQag9BmDsE8I0NsohBYgjQjhE7N2izHbOHtABecUIfwI0cecnaWtWE3PYdgtg3DIDQpHpfVK+YHQNSogjV6tMLFUW+dENwE8oQkUFF7CCMCOZ5HUZo+5E0anC1MHaS+0irB2WvtYCSgRiLZh-FJCROLkhAA */
